requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.11.1
datasets>=2.19.0
accelerate>=0.30.0
mlflow>=2.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
huggingface-hub>=0.23.0
safetensors>=0.4.0
tqdm
pyyaml
python-dotenv
bitsandbytes
matplotlib>=3.5.0
seaborn>=0.12.0
requests
psutil